backend/requirements.txt
# API and Web
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
requests==2.31.0
httpx==0.24.1

# Canvas LMS API
canvasapi>=2.0.0

# Data Processing
# Use a version of numpy that doesn't require compilation
numpy>=1.20.0
pandas>=1.3.0
pydantic==2.3.0
pydantic-settings==2.0.3

# Document Processing
python-docx==0.8.11
python-pptx==0.6.21
pdfplumber==0.10.2
pillow==10.0.1
pytesseract==0.3.10
pdf2image==1.16.3
pypdf==3.16.0
pypdfium2==4.25.0
python-dotenv==1.0.0

# Open Source OCR Engines
# Use binary distributions where possible
opencv-python>=4.8.0
unstructured>=0.10.0
# Comment out packages that might be difficult to install
# paddlepaddle==2.5.2
# paddleocr==2.6.0.3
# unstructured-inference==0.7.21

# AI and ML
langchain>=0.0.300
sentence-transformers>=2.2.0

# Database (Optional)
neo4j>=5.0.0

# Utilities
tqdm>=4.0.0
colorama>=0.4.0
python-dateutil>=2.8.0
